+++ src/main/java/roomescape/configuration/WebMvcConfiguration.java
@@ -6,7 +6,7 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
-import roomescape.controller.resolver.AccessTokenArgumentResolver;
+import roomescape.configuration.resolver.AccessTokenArgumentResolver;
 
 @Configuration
 public class WebMvcConfiguration implements WebMvcConfigurer {
