+++ src/main/java/roomescape/configuration/resolver/AccessTokenArgumentResolver.java
@@ -1,4 +1,4 @@
-package roomescape.controller.resolver;
+package roomescape.configuration.resolver;
 
 import java.util.Arrays;
 import java.util.Objects;
