+++ src/main/java/roomescape/controller/rest/ReservationController.java
@@ -12,7 +12,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import roomescape.controller.resolver.AccessToken;
+import roomescape.configuration.resolver.AccessToken;
 import roomescape.controller.rest.request.ReservationRequest;
 import roomescape.domain.Member;
 import roomescape.domain.Reservation;
